Give the BSP console API and internals explicit widths
The console API was written in bare int, which said nothing about what
fits. A read now returns int16_t, the narrowest type holding both a
0..255 byte and the negative BSP_CONSOLE_TIMEOUT sentinel; a write takes
the uint8_t it actually emits; printf and puts return int32_t, the count
or negative error the C library reports. char stays char for string data,
which the varargs contract requires.

Internally, counts the BSP owns become uint32_t: the byte count in
bsp_fpga.c's transaction(), the loop indices, the psram sizing locals,
and the fake console's queue, which holds only real bytes because the
timeout is produced on an empty queue rather than stored in it.

Three int and size_t declarations survive on purpose, at the boundaries
where the type is not ours to choose: getchar_timeout_us and
psram_reinitialize return the SDK's int, and psram_eid_to_size overrides
a weak SDK symbol whose size_t return type is fixed by the declaration
the SDK calls it through. Each converts explicitly rather than silently.

The image grows 8 bytes, from the sign extension the narrower console
types imply.

+++ firmware/src/application/application_console.c
@@ -66,10 +66,10 @@ static void stop_active_status(void) {
     console.status_mode = STATUS_DISABLED;
 }
 
-static void echo_character(int character) {
+static void echo_character(int16_t character) {
     if (!console.quiet && console.echo_enabled) {
         mark_write();
-        BSP_ConsolePutChar(character);
+        BSP_ConsolePutChar((uint8_t)character);
     }
 }
 
@@ -116,7 +116,7 @@ static void redraw_line(void) {
     }
 }
 
-static void process_character(int character) {
+static void process_character(int16_t character) {
     if (character == '\n' && console.swallow_lf) {
         console.swallow_lf = false;
         return;
@@ -167,7 +167,7 @@ void application_console_start(void) {
 
 void application_console_poll(void) {
     BSP_WatchdogMarkerSet(APPLICATION_DIAGNOSTICS_MARKER_CONSOLE_READ);
-    int character = BSP_ConsoleGetCharTimeoutUs(1000);
+    int16_t character = BSP_ConsoleGetCharTimeoutUs(1000);
     console.current_time_ms = BSP_TimeNowMs();
 
     if (character != BSP_CONSOLE_TIMEOUT) {
