+++ firmware/src/bsp/bsp_console.c
@@ -10,31 +10,31 @@ void BSP_ConsoleInit(void)
     stdio_init_all();
 }
 
-int BSP_ConsoleGetCharTimeoutUs(const uint32_t timeout_us)
+int16_t BSP_ConsoleGetCharTimeoutUs(const uint32_t timeout_us)
 {
     const int character = getchar_timeout_us(timeout_us);
     if (character == PICO_ERROR_TIMEOUT)
     {
         return BSP_CONSOLE_TIMEOUT;
     }
-    return character;
+    return (int16_t)character;
 }
 
-int BSP_ConsolePutChar(const int character)
+int16_t BSP_ConsolePutChar(const uint8_t character)
 {
-    return putchar(character);
+    return (int16_t)putchar(character);
 }
 
-int BSP_ConsolePrintf(const char *const format, ...)
+int32_t BSP_ConsolePrintf(const char *const format, ...)
 {
     va_list arguments;
     va_start(arguments, format);
     const int result = vprintf(format, arguments);
     va_end(arguments);
-    return result;
+    return (int32_t)result;
 }
 
-int BSP_ConsolePuts(const char *const text)
+int32_t BSP_ConsolePuts(const char *const text)
 {
-    return puts(text);
+    return (int32_t)puts(text);
 }
