+++ firmware/src/bsp/bsp_console.h
@@ -8,10 +8,10 @@
    confused with a real byte value, which is always in 0..255. */
 #define BSP_CONSOLE_TIMEOUT ((int16_t)-1)
 
-void BSP_ConsoleInit(void);
-int BSP_ConsoleGetCharTimeoutUs(const uint32_t timeout_us);
-int BSP_ConsolePutChar(const int character);
-int BSP_ConsolePrintf(const char *const format, ...);
-int BSP_ConsolePuts(const char *const text);
+void    BSP_ConsoleInit(void);
+int16_t BSP_ConsoleGetCharTimeoutUs(const uint32_t timeout_us);
+int16_t BSP_ConsolePutChar(const uint8_t character);
+int32_t BSP_ConsolePrintf(const char *const format, ...);
+int32_t BSP_ConsolePuts(const char *const text);
 
 #endif
